Created Item Spawn
Items can now be spawn according to a GameObject position. For some reasons it spawns with kinematic on, so it's floating for now. Not sure why thats going on.
Oh if you want to add a new object to work on both host and client you need to add "Client Network Transform" to any item. Then add the object you want into NetworkManager. Under "Network Prefabs".